refactor(ui): extract chat canonicalization ownership
